requirements.txt
streamlit>=1.34
pypdf>=4.2.0
Pillow>=10.3.0
pillow-heif>=0.19.0
reportlab>=4.2.0
pdf2image>=1.17.0
pymupdf>=1.24.0
pytesseract>=0.3.10
weasyprint>=62.3
python-docx>=1.1.2
openpyxl>=3.1.5
python-pptx>=1.0.2
rapidfuzz>=3.9.0
